# docs/swagger.yaml
definitions:
  handler.CreateCattleRequest:
    properties:
      earring:
        type: string
      gender:
        type: string
      name:
        type: string
    type: object
  handler.CreateCattleResponse:
    properties:
      data:
        $ref: '#/definitions/schemas.CattleResponse'
      message:
        type: string
    type: object
  handler.DeleteCattleResponse:
    properties:
      data:
        $ref: '#/definitions/schemas.CattleResponse'
      message:
        type: string
    type: object
  handler.ErrorResponse:
    properties:
      errorCode:
        type: string
      message:
        type: string
    type: object
  schemas.CattleResponse:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      earring:
        type: string
      gender:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
info:
  contact: {}
paths:
  /cattle:
    delete:
      consumes:
      - application/json
      description: Delete a cattle
      parameters:
      - description: Cattle identification
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.DeleteCattleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete cattle
      tags:
      - Cattle
    post:
      consumes:
      - application/json
      description: Create a new cattle
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateCattleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CreateCattleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create cattle
      tags:
      - Cattle
swagger: "2.0"